Extract MCP approval templates from core
Move the consequential MCP tool approval template JSON and renderer into a small resource crate, leaving codex-core with its existing internal wrapper path for callers.
