+++ codex-rs/mcp-tool-approval-templates/BUILD.bazel
@@ -1,0 +1,7 @@
+load("//:defs.bzl", "codex_rust_crate")
+
+codex_rust_crate(
+    name = "mcp-tool-approval-templates",
+    crate_name = "codex_mcp_tool_approval_templates",
+    compile_data = ["src/consequential_tool_message_templates.json"],
+)
