+++ codex-rs/mcp-tool-approval-templates/Cargo.toml
@@ -1,0 +1,21 @@
+[package]
+edition.workspace = true
+license.workspace = true
+name = "codex-mcp-tool-approval-templates"
+version.workspace = true
+
+[lib]
+doctest = false
+name = "codex_mcp_tool_approval_templates"
+path = "src/lib.rs"
+
+[lints]
+workspace = true
+
+[dependencies]
+serde = { workspace = true, features = ["derive"] }
+serde_json = { workspace = true }
+tracing = { workspace = true, features = ["log"] }
+
+[dev-dependencies]
+pretty_assertions = { workspace = true }
