+++ codex-rs/mcp-tool-approval-templates/src/consequential_tool_message_templates.json
